backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
tomli==2.0.1
websockets==12.0

# Optional for real hardware
RPi.GPIO==0.7.0  # Only needed on Raspberry Pi
adafruit-circuitpython-ads1x55==1.0.16  # For ADS1256
adafruit-circuitpython-max31855==3.3.3  # For MAX31855 thermocouple

# Development
pytest==7.4.3
black==23.12.0
mypy==1.7.1
